quiz-api/requirements.txt
autopep8==2.3.2
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycodestyle==2.14.0
PyJWT==2.5.0
python-dotenv==1.1.1
Werkzeug==3.1.3
gunicorn==20.1.0
pytest==7.4.0
python-dotenv==1.1.1
pytest-mock
